[ refactor ] Auth white list
