+++ src/main/java/greeny/backend/config/security/SecurityConfig.java
@@ -16,6 +16,8 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.filter.CorsFilter;
 
+import static org.springframework.http.HttpMethod.*;
+
 @Configuration
 @RequiredArgsConstructor
 public class SecurityConfig {
@@ -25,7 +27,7 @@ public class SecurityConfig {
     private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
     private final JwtAccessDeniedHandler jwtAccessDeniedHandler;
 
-    private static final String[] AUTH_WHITELIST_WITH_MEMBER_AUTH = {
+    private static final String[] AUTH_WHITELIST_WITH_MEMBER_AUTH = {  // 사용자 인증 white list
             "/swagger-ui/**",
             "/api-docs/**",
             "/api",
@@ -35,15 +37,17 @@ public class SecurityConfig {
             "/api/auth/password",
             "/api/auth/reissue"
     };
-
-    private static final String[] AUTH_WHITELIST_WITH_GET_METHOD = {
+    private static final String[] AUTH_WHITELIST_WITH_ECO_GET_METHOD = {  // Store or Product white list
             "/api/stores/simple",
             "/api/stores",
             "/api/products/simple",
-            "/api/products",
+            "/api/products"
+    };
+    private static final String[] AUTH_WHITELIST_WITH_REVIEW_GET_METHOD = {  // Review white list
             "/api/reviews/all",
-            "/api/reviews/simple",
-            "/api/reviews",
+            "/api/reviews/simple"
+    };
+    private static final String[] AUTH_WHITELIST_WITH_COMMUNITY_GET_METHOD = {  // Community white list
             "/api/posts/**",
             "/api/comments/**"
     };
@@ -73,7 +77,11 @@ public class SecurityConfig {
                 .authorizeRequests(authorize -> authorize
                         .antMatchers(AUTH_WHITELIST_WITH_MEMBER_AUTH)
                         .permitAll()
-                        .antMatchers(HttpMethod.GET, AUTH_WHITELIST_WITH_GET_METHOD)  // 인증 없이 조회 가능한 API 목록
+                        .antMatchers(GET, AUTH_WHITELIST_WITH_ECO_GET_METHOD)
+                        .permitAll()
+                        .antMatchers(GET, AUTH_WHITELIST_WITH_REVIEW_GET_METHOD)
+                        .permitAll()
+                        .antMatchers(GET, AUTH_WHITELIST_WITH_COMMUNITY_GET_METHOD)
                         .permitAll()
                         .anyRequest()
                         .authenticated()
