+++ src/main/java/greeny/backend/domain/community/controller/CommentController.java
@@ -1,7 +1,7 @@
-package greeny.backend.domain.board.controller;
+package greeny.backend.domain.community.controller;
 
-import greeny.backend.domain.board.dto.WriteCommentRequestDto;
-import greeny.backend.domain.board.service.CommentService;
+import greeny.backend.domain.community.dto.WriteCommentRequestDto;
+import greeny.backend.domain.community.service.CommentService;
 import greeny.backend.domain.member.service.MemberService;
 import greeny.backend.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
