+++ src/main/java/greeny/backend/domain/community/dto/WriteCommentRequestDto.java
@@ -1,7 +1,7 @@
-package greeny.backend.domain.board.dto;
+package greeny.backend.domain.community.dto;
 
-import greeny.backend.domain.board.entity.Comment;
-import greeny.backend.domain.board.entity.Post;
+import greeny.backend.domain.community.entity.Comment;
+import greeny.backend.domain.community.entity.Post;
 import greeny.backend.domain.member.entity.Member;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
