+++ src/main/java/greeny/backend/domain/community/repository/PostRepository.java
@@ -1,10 +1,9 @@
-package greeny.backend.domain.board.repository;
+package greeny.backend.domain.community.repository;
 
-import greeny.backend.domain.board.entity.Post;
+import greeny.backend.domain.community.entity.Post;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
