+++ src/main/java/greeny/backend/domain/member/entity/Member.java
@@ -1,8 +1,8 @@
 package greeny.backend.domain.member.entity;
 
 import greeny.backend.domain.AuditEntity;
-import greeny.backend.domain.board.entity.Comment;
-import greeny.backend.domain.board.entity.Post;
+import greeny.backend.domain.community.entity.Comment;
+import greeny.backend.domain.community.entity.Post;
 import greeny.backend.domain.bookmark.entity.ProductBookmark;
 import greeny.backend.domain.bookmark.entity.StoreBookmark;
 import greeny.backend.domain.review.entity.ProductReview;
@@ -18,7 +18,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static javax.persistence.CascadeType.*;
-import static javax.persistence.FetchType.*;
 
 @Entity
 @Getter
